Subsystem lifecycle refactor
This should hopefully be the last refactor of the subsystem architecture
for now.  The ILifecycle interface now stores more specific information
on the unit's current state, which allows for more error checking.
Additionally, the AbstractSubsys class now provides the onInit() and
onExit() callbacks to its children, which are called in the init() and
exit() methods respectively.  This allows for better error checking and
updating the state accordingly.