Add auth subsystem
Part of a series on #2.
This version includes a kind of nasty hack to provide backwards
compatibility with the old JWT wrapper in src/util/jwt.ts (by far not as
nasty as the original JWT utils, though).  That will get removed once I
finally replace the routing stuff with an own subsystem, which turns
out to be more difficult than expected.