Change default certificate format to PEM
I don't know if I should be mad at myself because I didn't read the JWT
specs in their entirety, or at the guys behind the jsonwebtoken package
for their crappy documentation.  Anyways, jsonwebtoken only wants PEMs
and the 'secretOrPrivateKey' parameter is always interpreted as a
certificate when the algorithm is ES*** (kind of obvious in hindsight,
but this still should have been mentioned _somewhere_ in the docs).