Rework middleware system so that it actually works
Turns out I had a little misunderstanding about how decorators are
evaluated.  Factories are only called once _per class_ and then reused
for all instances, which meant the injectValue decorator always set the
exact same middleware object (not a copy of it) for all instances of
AbstractRoute.  This ultimately meant that middleware added in one
object was there in _all_ objects, which kind of misses the point of
havin a decorator in the first place.