Replace old routers with new architecture
Part of a series on #2.
This should be one of the last commits to fix this issue; that took me
long enough.  The next step is to get rid of all those legacy code
snippets to remain compatible with the old architecture, which will
probably take the rest of the day.  Really looking forward to that.