Get rid of Express namespace pollution
Part of a series on #2 (we're almost there!).
Maybe it's a good idea to just get rid of Express all together, I mean
I'm using only like 20% of its features and implementing an own routing
framework might even be faster.  But that of course is no priority right
now, maybe just before the 1.0 release in like, 5 years from now.